Create spacy_parse_authors
This is a very tentative attempt to parse the dct:creator value for each Document. This is a hack and almost all of this needs to be redone. I (MD) created this when first learning Spacy but he didn't read far enough to realize there actually already are pattern matchers for names and the current patterns work for much of the data but not all. Currently (in order to get enough data for Nivedita's journal article) I edited some of the names to make them fit these simple patterns So at some point we need to delete all the current values for hasAuthor and redo them.