Delete drmo_rd_4_5_24.ttl
In the latest version I used a SPARQL query to delete duplicate documents (there are a lot for some reason) but I forgot that the way SPARQL works it ended up deleting both copies of the duplicate. So I deleted that ontology. On monday i'll fix my SPARQL transformation and rerun it and also redo the other changes I made. For now the 4_4 version is the latest one.