Updated Ontology and SPARQL Utilities
Deleted (again) all the puns and the skos:Concept class. Deleting puns also deleted labels for punned classes, which was a pain but an example of why puns are not a good idea. Think I added all the labels back, most with the SPARQL utility some (e.g.. OBO) manually.